requirements.txt
streamlit>=1.25
opencv-python
numpy
pillow
streamlit-drawable-canvas
